+++ next.config.js
@@ -4,6 +4,22 @@ const nextConfig = {
     experimental: {
         serverComponentsExternalPackages: ['prisma', '@prisma/client'],
     },
+    images: {
+        remotePatterns: [
+            {
+                protocol: 'https',
+                hostname: 'avatars.githubusercontent.com',
+            },
+            {
+                protocol: 'https',
+                hostname: 'lh3.googleusercontent.com',
+            },
+            {
+                protocol: 'https',
+                hostname: '*.googleusercontent.com',
+            },
+        ],
+    },
 };
 
 module.exports = nextConfig;